Move crash testing to CrashHelpers.h
