+++ interface/src/main.cpp
@@ -33,84 +33,7 @@
 #include <CrashReporter.h>
 #endif
 
-#include <csignal>
-
-namespace crash {
-
-    void pureVirtualCall() {
-        struct B {
-            B() {
-                qDebug() << "Pure Virtual Function Call crash!";
-                Bar();
-            }
-
-            virtual void Foo() = 0;
-
-            void Bar() {
-                Foo();
-            }
-        };
-
-        struct D : public B {
-            void Foo() {
-                qDebug() << "D:Foo()";
-            }
-        };
-
-        B* b = new D;
-        qDebug() << "About to make a pure virtual call";
-        b->Foo();
-    }
-        
-    void doubleFree() {
-        qDebug() << "About to double delete memory";
-        int* blah = new int(200);
-        delete blah;
-        delete blah;
-    }
-
-    void nullDeref() {
-        qDebug() << "About to dereference a null pointer";
-        int* p = nullptr;
-        *p = 1;
-    }
-
-    void doAbort() {
-        qDebug() << "About to abort";
-        abort();
-    }
-
-    void outOfBoundsVectorCrash() {
-        qDebug() << "std::vector out of bounds crash!";
-        std::vector<int> v;
-        v[0] = 5;
-    }
-
-    void newFault() {
-        qDebug() << "About to crash inside new fault";
-        // Force crash with large allocation
-        int *pi = new int[std::numeric_limits<uint64_t>::max()];
-    }
-}
-//
-//LONG WINAPI TopLevelExceptionHandler(PEXCEPTION_POINTERS pExceptionInfo) {
-//    qDebug() << "Got exception";
-//
-//    mpSender.unhandledExceptionHandler(pExceptionInfo);
-//
-//    return EXCEPTION_CONTINUE_SEARCH;
-//}
-
-
-#if HAS_BUGSPLAT
-    // Prevent other threads from hijacking the Exception filter, and allocate 4MB up-front that may be useful in
-    // low-memory scenarios.
-//    static const DWORD BUG_SPLAT_FLAGS = MDSF_PREVENTHIJACKING | MDSF_USEGUARDMEMORY;// | MDSF_CUSTOMEXCEPTIONFILTER;
-//    static const char* BUG_SPLAT_DATABASE = "interface_alpha";
-//    static const char* BUG_SPLAT_APPLICATION_NAME = "Interface";
-//    static MiniDmpSender mpSender { BUG_SPLAT_DATABASE, BUG_SPLAT_APPLICATION_NAME, qPrintable(BuildInfo::VERSION),
-//                                    nullptr, BUG_SPLAT_FLAGS };
-#endif
+#include <CrashHelpers.h>
 
 int main(int argc, const char* argv[]) {
 #if HAS_BUGSPLAT
